app/requirements.txt
Flask
psutil
python-dotenv
